feat: add ERC721 tests for BOL
Simplified the tests by just verifying they assert the correct
interface. Under the hood we're just using the openzeppelin library.
If we wanted to be truly testing we would have to confirm functionality.
Instead the easy way out is to call the `supportsInterface` function and
naively believe that the interface truly is implemented.